requirements.txt
# Dependencias del proyecto PLS Biomédico
# Python 3.8+

# ===============================================
# CORE ML - PyTorch y Transformers
# ===============================================
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
sentencepiece>=0.1.99

# ===============================================
# PEFT - Parameter Efficient Fine-Tuning
# ===============================================
peft>=0.4.0
bitsandbytes>=0.39.0; sys_platform != 'win32'  # No disponible en Windows

# ===============================================
# DATA PROCESSING
# ===============================================
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# ===============================================
# NLP METRICS AND TOOLS
# ===============================================
rouge-score>=0.1.2
bert-score>=0.3.13
nltk>=3.8
textstat>=0.7.3
evaluate>=0.4.0

# ===============================================
# VERSIONING AND REPRODUCIBILITY
# ===============================================
dvc[s3]>=3.0.0
boto3>=1.26.0
s3fs>=2023.1.0

# ===============================================
# VISUALIZATION AND DASHBOARDS
# ===============================================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
streamlit>=1.24.0
gradio>=3.35.0

# ===============================================
# JUPYTER NOTEBOOKS
# ===============================================
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0

# ===============================================
# UTILITIES
# ===============================================
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0
click>=8.1.0
colorama>=0.4.6

# ===============================================
# DEVELOPMENT (Optional - descomenta si necesitas)
# ===============================================
# black>=23.0.0
# flake8>=6.0.0
# pytest>=7.4.0
# mypy>=1.4.0
